deleted some comments, fixed tests, moved tuflow_to_dat to separate file, added XML prefix to 2D cc
